Derive Macro oak_derive::OakExports

#[derive(OakExports)]

Implements the necessary bindings to make the annotated struct act as a gRPC-processing Oak Node.

May only be used on struct objects. Assumes that the default pre-defined port names ("grpc_in") is used to identify the gRPC input channel.

At most one struct may be annotated with this, as it produces global symbols that would otherwise conflict if implemented multiple times.

extern crate oak;
extern crate protobuf;
use oak::grpc::OakNode;
use protobuf::ProtobufEnum;

#[derive(oak_derive::OakExports)]
struct Node;

impl OakNode for Node {
    fn new() -> Self { Node }
    fn invoke(&mut self, method: &str, req: &[u8], writer: oak::grpc::ChannelResponseWriter) { /* ... */ }
}